% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdb-bootstrap-generic.R
\name{bootstrap.ecdb}
\alias{bootstrap.ecdb}
\alias{bootstrap}
\alias{bootstrap,ecdb-method}
\title{Bootstrap data for the Elliptic DB (ECDB)}
\usage{
\method{bootstrap}{ecdb}(object, action = "all", skip.existing = TRUE)

bootstrap(object, action = "all", skip.existing = TRUE)

\S4method{bootstrap}{ecdb}(object, action = "all", skip.existing = TRUE)
}
\arguments{
\item{object}{an object of ecdb class.}

\item{action}{the action operating on the ecdb.}

\item{skip.existing}{logical, if \code{TRUE} (default), skip if action already done in history.}
}
\value{
Row count.
}
\description{
Main interface to generate data for ECDB based on the configuration.
}
\keyword{ecdb}
