% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-discr-generic.R
\name{discr.ecd}
\alias{discr.ecd}
\alias{discr}
\alias{discr,ecd-method}
\title{Discriminant of the elliptic curve \eqn{y(x)}}
\usage{
\method{discr}{ecd}(object, no.validate = FALSE)

discr(object, no.validate = FALSE)

\S4method{discr}{ecd}(object, no.validate = FALSE)
}
\arguments{
\item{object}{an object of ecd class}

\item{no.validate}{logical, if \code{TRUE}, don't validate presence of \code{beta}.
Default is \code{FALSE}.}
}
\value{
the discriminant
}
\description{
Discriminant of the elliptic curve \eqn{y(x)}
}
\examples{
d <- ecd(-1,1)
discr(d)

}
\author{
Stephen H-T. Lihn
}
\keyword{stats}
