% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-package.R
\docType{package}
\name{ecd-package}
\alias{ecd-package}
\title{ecd: A package for the stable lambda distribution family.}
\description{
The ecd package provides the core classes and functions for the stable lambda distribution family.
The stable lambda distribution is implemented in \code{\link{dsl}} section.
The lambda distribution uses the \code{ecld} namespace. SGED is considered part of ecld.
(See \code{\link{ecld-class}} for definition.)
The original elliptic lambda distribution uses the generic methods or \code{ecd} namespace. 
(See \code{\link{ecd-class}} for definition.)
The option pricing API uses the \code{ecop} namespace.
(See \code{\link{ecop-class}} for definition.)
Most helper utilities are named under either \code{ecd} or \code{ecld}.
}
\seealso{
The two main classes are \code{\link{ecd-class}} and \code{\link{ecld-class}}
}
\author{
Stephen H-T. Lihn
}
