% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-adj-gamma-method.R
\name{ecd.adj_gamma}
\alias{ecd.adj2gamma}
\alias{ecd.adj_gamma}
\title{Discriminant-adjusted gamma}
\usage{
ecd.adj_gamma(gamma)

ecd.adj2gamma(adj_gamma)
}
\arguments{
\item{gamma}{numeric, the gamma paramter}

\item{adj_gamma}{numeric, the disciminant-adjusted gamma}
}
\value{
adjusted gamma (or the reverse of adjustment)
}
\description{
Adjust gamma by diescriminant conversion formula so that
the critical line is a straight 45-degree line. 
The inverse adjustment is also provided.
}
\examples{
gamma2 <- ecd.adj_gamma(c(1,2))
gamma <- ecd.adj2gamma(c(1,2))
}
\keyword{discriminant}

