% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-cubic-method.R
\name{ecd.cubic}
\alias{ecd.cubic}
\title{Generate or solve the cubic polynomial for ecd}
\usage{
ecd.cubic(object, x = 0, solve = TRUE)
}
\arguments{
\item{object}{An object of ecd class}

\item{x}{A vector of x dimension}

\item{solve}{Logical, solve the polynomial, default = TRUE.}
}
\value{
list of the polynomial object, or result of solve.
}
\description{
Generate or solve the polynomial from ecd. 
This is usually transient for solve.
Or it can be used for studying singular points.
}
\examples{
d <- ecd()
ecd.cubic(d)
ecd.cubic(d, 0)
}
\keyword{cubic}

