% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-cusp-a2r-method.R
\name{ecd.cusp_a2r}
\alias{ecd.cusp_a2r}
\alias{ecd.cusp_r2a}
\title{Conversion between alpha and gamma for cusp distribution}
\usage{
ecd.cusp_a2r(alpha)

ecd.cusp_r2a(gamma)
}
\arguments{
\item{alpha}{numeric}

\item{gamma}{numeric}
}
\value{
gamma for \code{a2r}; alpha for \code{r2a}.
}
\description{
\code{ecd.cusp_a2r} converts from alpha to gamma.
\code{ecd.cusp_r2a} converts from gamma to alpha.
}
\examples{
gamma <- ecd.cusp_a2r(alpha=1)
alpha <- ecd.cusp_r2a(gamma=1)
}
\author{
Stephen H-T. Lihn
}
\keyword{cusp}
