% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-data-method.R
\name{ecd.data}
\alias{ecd.data}
\title{Read sample data}
\usage{
ecd.data(symbol = "dji")
}
\arguments{
\item{symbol}{Character for the symbol of the time series. Default: dji}
}
\value{
The xts object for the time series
}
\description{
Read sample data into xts by specifying the symbol.
The xts object has two rows: the prices indexed by the dates.
}
\examples{
dji <- ecd.data()
wti <- ecd.data("wti")
}
\keyword{sample-data}
\keyword{timeseries}
\keyword{xts}

