% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-data-stats-method.R
\name{ecd.data_stats}
\alias{ecd.data_stats}
\title{Statistics and histogram on log returns}
\usage{
ecd.data_stats(
  ts = "dji",
  breaks = 20,
  merge_tails = c(0, 0),
  with.tail = FALSE,
  tail.N1 = 7,
  tail.N2 = 5
)
}
\arguments{
\item{ts}{can be either a symbol of sample data,
or the xts object from sample data}

\item{breaks}{A length-one numeric, breaks for generating the histogram.}

\item{merge_tails}{A length-two numeric vector.
The first element is how many points in the left tail of histogram to be dropped during fitting.
The second element is how many points in the right tail of histogram to be dropped during fitting.}

\item{with.tail}{logical, include tail statistics, mainly on asypmtotic kurtosis. Default: \code{FALSE}.}

\item{tail.N1}{a numeric, defining the wider range of tail statistics}

\item{tail.N2}{a numeric, defining the smaller range of tail statistics}
}
\value{
The xts object containing ecd added attributes
}
\description{
Statistics and histogram on log returns are added to the xts attributes
}
\examples{
dji <- ecd.data_stats(ecd.data("dji"))
dji <- ecd.data_stats("dji")
}
\keyword{sample-data}
\keyword{statistics}
