% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-diff-method.R
\name{ecd.diff}
\alias{ecd.diff}
\title{Utility to diff a vector of numeric or mpfr to get first derivative}
\usage{
ecd.diff(y, x, pad = 0)
}
\arguments{
\item{y}{a vector of numeric or mpfr}

\item{x}{a vector of numeric or mpfr}

\item{pad}{integer, to manage padding so that the output vector has
the same length as the input. 0 for no padding,
1 to repeat the first element, -1 to repeat the last element.}
}
\value{
the derivative vector
}
\description{
This utility uses diff to get first derivative dy/dx.
but it handles mpfr vector properly
}
\examples{
d <- ecd.diff(c(10,20,30), c(1,2,3), pad=1)

}
\keyword{utility}
