% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-estimate-const-method.R
\name{ecd.estimate_const}
\alias{ecd.estimate_const}
\title{Estimate the normalization constant for an ecd object}
\usage{
ecd.estimate_const(object)
}
\arguments{
\item{object}{An object of ecd class}
}
\value{
numeric, estimated const
}
\description{
This is an internal helper function for ecd constructor
Its main function is to estimate \code{const} using analytical formula,
without any dependency on statistics and numerical integration.
}
\examples{
ecd.estimate_const(ecd(100,100, sigma=0.1, bare.bone=TRUE))

}
\keyword{constructor}

