% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-has-quantile-method.R
\name{ecd.has_quantile}
\alias{ecd.has_quantile}
\title{Whether the ecd object has quantile data or not}
\usage{
ecd.has_quantile(object)
}
\arguments{
\item{object}{an object of ecd class}
}
\value{
logical, whether the object has quantile data or not.
}
\description{
Whether the ecd object has quantile data or not. This is mostly for internal use.
}
\author{
Stephen H-T. Lihn
}
\keyword{distribution}
