% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-lag-method.R
\name{ecd.lag}
\alias{ecd.lag}
\title{Utility to shift a vector of numeric or mpfr}
\usage{
ecd.lag(x, shift = 1, na.omit = FALSE)
}
\arguments{
\item{x}{a vector of numeric or mpfr}

\item{shift}{integer, cells to shift}

\item{na.omit}{logical, whether to remove the NAs}
}
\value{
the shifted vector
}
\description{
This utility is basically the same as \code{Hmisc::Lag},
but it handles mpfr vector properly.
}
\examples{
x <- ecd.lag(c(1,2,3))
y <- ecd.lag(ecd.mpfr(c(1,2,3)))
}
\keyword{utility}

