% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-manage-hist-tails-method.R
\name{ecd.manage_hist_tails}
\alias{ecd.manage_hist_tails}
\title{Manage histogram tails}
\usage{
ecd.manage_hist_tails(htu, merge_tails = c(0, 0))
}
\arguments{
\item{htu}{list, input histuple}

\item{merge_tails}{length-two numeric vector, points to be merged for left and right tails}
}
\value{
list, histuple
}
\description{
Manage histogram tails to remove very far outliers.
\code{histuple} is \code{list(hx = hist\$mids, hy = hist\$counts)},
which is an internal representation of histogram
}
\examples{
\dontrun{
htu2 <- ecd.manage_hist_tails(htu, c(1,2))
}
}
\author{
Stephen H-T. Lihn
}
\keyword{data}

