% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-max-kurtosis-method.R
\name{ecd.max_kurtosis}
\alias{ecd.max_kurtosis}
\title{Utility to calculate where the maximum kurtosis is on the positive j=0 line}
\usage{
ecd.max_kurtosis(jinv = 0)
}
\arguments{
\item{jinv}{specify 0 (default) or 1728.}
}
\value{
numeric vector, in which the first element is alpha, 
        and the second element is the maximum kurtosis.
}
\description{
This utility calculates the kurtosis for alpha from 2.85 to 3.00. Then
the location and value of maximum kurtosis is presented.
}
\examples{
\dontrun{
    k <- ecd.max_kurtosis()
    alpha <- k[1]
    kurtosis <- k[2]
}
}
\author{
Stephen H-T. Lihn
}
\keyword{utility}
