% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-mp2f-method.R
\name{ecd.mp2f}
\alias{ecd.mp2f}
\title{Wrapper to convert mpfr to numeric}
\usage{
ecd.mp2f(x)
}
\arguments{
\item{x}{an object of \code{mpfr} class. If \code{x} is numeric class,
it will be passed through.}
}
\value{
a numeric vector
}
\description{
Convert mpfr to numeric primarily for display messages.
}
\examples{
x <- ecd.mp2f(ecd.mpfr(c(1,2,3)))
}
\keyword{utility}
