% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-mpfr-method.R
\docType{data}
\name{ecd.mpfr}
\alias{ecd.devel}
\alias{ecd.erf}
\alias{ecd.mp1}
\alias{ecd.mpfr}
\title{Wrapper to convert numeric to mpfr}
\format{An object of class \code{mpfr} of length 1.}
\usage{
ecd.mpfr(x, precBits = getOption("ecd.precBits"))

ecd.mp1

ecd.erf(x)

ecd.devel()
}
\arguments{
\item{x}{a numeric vector or list. If \code{x} is mpfr class,
it will be passed through.}

\item{precBits}{an integer for mpfr precBits.
Default is from \code{getOption("ecd.precBits")}.}
}
\value{
The \code{mpfr} object
}
\description{
Convert numeric to mpfr for ecd calculations.
\code{ecd.mp1} is the constant 1 wrapped in mpfr class.
\code{ecd.erf} is a wrapper on \code{Rmpfr::erf}.
\code{ecd.devel} is a developer tool to size down intensive mpfr tests for CRAN. Set \code{ecd_devel} in R options or OS env to change its value.
}
\examples{
x <- ecd.mpfr(1)
y <- ecd.mpfr(c(1,2,3))
z <- ecd.mp1
}
\keyword{datasets}
\keyword{utility}

