% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-mpnum-method.R
\name{ecd.mpnum}
\alias{ecd.mpnum}
\alias{ecd.ifelse}
\alias{ecd.sapply}
\alias{ecd.mcsapply}
\title{Wrappers for ecd to maintain consistent type between mpfr and numeric}
\usage{
ecd.mpnum(object, x)

ecd.ifelse(object, test, yes, no)

ecd.sapply(object, x, FUN, ...)

ecd.mcsapply(object, x, FUN, ...)
}
\arguments{
\item{object}{an object of ecd class}

\item{x}{a vector of numeric or mpfr.}

\item{test}{logical, test of \code{ifelse}.}

\item{yes}{return values for true elements of test}

\item{no}{return values for false elements of test}

\item{FUN}{the function to be applied to each element of x}

\item{...}{optional arguments to \code{FUN}}
}
\value{
a numeric or mpfr vector
}
\description{
Primarily to make sure \code{x} is converted to mpfr vector if it is not,
when \code{use.mpfr} is set.
}
\author{
Stephen H. Lihn
}
\keyword{utility}
