% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-pdf-method.R
\name{ecd.pdf}
\alias{ecd.pdf}
\title{Calculate the PDF of an ecd object}
\usage{
ecd.pdf(object, x)
}
\arguments{
\item{object}{an object of ecd class}

\item{x}{numeric vector of \eqn{x} dimension}
}
\value{
numeric vector of the PDF
}
\description{
Calculate the PDF of an ecd object
}
\examples{
d <- ecd()
x <- seq(-10, 10, by=1)
ecd.pdf(d,x)
}
\author{
Stephen H-T. Lihn
}
\keyword{distribution}
\keyword{pdf}

