% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-read-csv-by-symbol-method.R
\name{ecd.read_csv_by_symbol}
\alias{ecd.read_csv_by_symbol}
\title{Read csv file of sample data}
\usage{
ecd.read_csv_by_symbol(symbol = "dji")
}
\arguments{
\item{symbol}{Character for the symbol of the time series. Default: dji}
}
\value{
The data.frame object
}
\description{
This is a helper utility to read sample csv file into data frame.
The main use for external users is to read the option data
since it has a different format than other price timeseries data.
}
\examples{
dji <- ecd.read_csv_by_symbol("dji")
spx <- ecd.read_csv_by_symbol("spxoption2")

}
\keyword{sample-data}
\keyword{timeseries}

