% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-read-symbol-conf-method.R
\name{ecd.read_symbol_conf}
\alias{ecd.read_symbol_conf}
\title{Read conf for sample data}
\usage{
ecd.read_symbol_conf(symbol, conf_file = "conf/ecd-fit-conf.yml")
}
\arguments{
\item{symbol}{Character. The symbol of sample data. Default: dji.}

\item{conf_file}{File name fof symbol config, default to conf/ecd-fit-conf.yml}
}
\value{
the conf object
}
\description{
Read conf for sample data
}
\examples{
\dontrun{
conf <- ecd.read_symbol_conf("dji")
}
}
\keyword{fit}
\keyword{sample-data}
