% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-sd-method.R
\name{ecd.sd}
\alias{ecd.kurt}
\alias{ecd.kurtosis}
\alias{ecd.mean}
\alias{ecd.sd}
\alias{ecd.skewness}
\alias{ecd.var}
\title{Standard deviation, variance, mean, skewness, and kurtosis of ecd}
\usage{
ecd.sd(object)

ecd.var(object)

ecd.mean(object)

ecd.skewness(object)

ecd.kurt(object)

ecd.kurtosis(object)
}
\arguments{
\item{object}{an object of ecd class}
}
\value{
numeric or mpfr
}
\description{
Convenience wrappers around ecd's stats data
}
\examples{
d <- ecd(-1,1)
ecd.sd(d)
ecd.var(d)
ecd.mean(d)
ecd.skewness(d)
ecd.kurt(d)

}
\keyword{stats}

