% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-setup-const-method.R
\name{ecd.setup_const}
\alias{ecd.setup_const}
\title{Integration preprocessor for an ecd object}
\usage{
ecd.setup_const(object, verbose = FALSE)
}
\arguments{
\item{object}{An object of ecd class}

\item{verbose}{logical, display timing information, for debugging purpose.}
}
\value{
\code{list(const, const_left_x, const_right_x)}
}
\description{
This is an internal helper function for ecd constructor
Its main function is to determine \code{const}, \code{const_left_x}, and \code{const_right_x} during object construction.
}
\examples{
ecd.toString(ecd(-1,1, sigma=0.1))

}
\author{
Stephen H. Lihn
}
\keyword{constructor}

