% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-toString-method.R
\name{ecd.toString}
\alias{ecd.toString}
\title{String representation of ecd}
\usage{
ecd.toString(object, full = FALSE)
}
\arguments{
\item{object}{An object of ecd class}

\item{full}{logical, indicating if long form (multiple lines) should be rendered.}
}
\value{
character
}
\description{
A string representation of an ecd object. Can be used for warning or error.
}
\examples{
ecd.toString(ecd(-1,1, sigma=0.1))

}
\keyword{constructor}
