% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-uniroot-method.R
\name{ecd.uniroot}
\alias{ecd.uniroot}
\title{Uniroot wrapper}
\usage{
ecd.uniroot(f, lower, upper, use.mpfr = FALSE,
  tol = .Machine$double.eps^0.25, maxiter = 100)
}
\arguments{
\item{f}{the function for which the root is sought.}

\item{lower, upper}{the lower and upper end points of
the interval to be searched.}

\item{use.mpfr}{logical, to use MPFR (default), or else uniroot in stats.}

\item{tol}{the desired accuracy (convergence tolerance).}

\item{maxiter}{the maximum number of iterations.}
}
\value{
uniroot result
}
\description{
This function wraps ordinary uniroot and unirootR (from Rmpfr)
to the same interface.
}
\author{
Stephen H. Lihn
}
\keyword{utility}

