% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-y0-isomorphic-method.R
\name{ecd.y0_isomorphic}
\alias{ecd.y0_isomorphic}
\title{The analytic solution of \eqn{y(0)} via isomorphic mapping.}
\usage{
ecd.y0_isomorphic(theta = NaN, R = 1, object = NULL)
}
\arguments{
\item{theta}{numeric vector, the polar coordinate}

\item{R}{numeric vector, the polar coordinate}

\item{object}{optionally, a single ecd object}
}
\value{
the value of y(0)
}
\description{
This utility can be called two ways: 
(a) specify R and theta; (b) provide the ecd object.
But not at the same time.
}
\examples{
t <- 45/180*pi
ecd.y0_isomorphic(t)

}
\keyword{solve}

