% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdattr-enrich-method.R
\name{ecdattr.enrich}
\alias{ecdattr.enrich}
\title{Enrich a basic \code{ecdattr} object}
\usage{
ecdattr.enrich(p)
}
\arguments{
\item{p}{a basic \code{ecdattr} object}
}
\value{
an enriched \code{ecdattr} object
}
\description{
It takes a basic \code{ecdattr} object, enrich it with ecd attributes.
This function is computationally heavy. So the objects are often wrapped in a list 
and computed via \code{parallel::mclapply}.
}
\keyword{ecdattr}
