% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdattr-pairs-method.R
\name{ecdattr.pairs}
\alias{ecdattr.pairs}
\title{Create a list of basic \code{ecdattr} objects}
\usage{
ecdattr.pairs(alpha, gamma, cusp = 0, use.mpfr = FALSE)
}
\arguments{
\item{alpha, gamma}{numeric vectors}

\item{cusp}{numeric, representing type of cusp. Only 0 (default) and 1 are allowed.}

\item{use.mpfr}{logical, whether to use mpfr for ecd object, default is \code{FALSE}.}
}
\value{
a list of basic \code{ecdattr} objects.
}
\description{
The list is created by the Cartesian product between \code{alpha} and \code{gamma}.
This contains the data points of a rectangular area defined by \code{alpha,gamma}.
If \code{cusp} is 1, data points are on the critical line specified by \code{alpha}.
}
\keyword{ecdattr}
