% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdattr-pairs-polar-method.R
\name{ecdattr.pairs_polar}
\alias{ecdattr.pairs_polar}
\title{Create a list of basic \code{ecdattr} objects in polar coordinate}
\usage{
ecdattr.pairs_polar(R, theta, cusp = 0, use.mpfr = FALSE)
}
\arguments{
\item{R, theta}{numeric vectors}

\item{cusp}{numeric, representing type of cusp. Only 0 (default) and 1 are allowed.}

\item{use.mpfr}{logical, whether to use mpfr for ecd object, default is \code{FALSE}.}
}
\value{
a list of basic \code{ecdattr} objects.
}
\description{
The list is created by the Cartesian product between \code{R} and \code{theta}.
This contains the data points of a circular area defined by \code{R,theta}.
If \code{cusp} is 1, data points are on the critical line specified by \code{R}.
}
\keyword{ecdattr}

