% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdb-class.R
\docType{class}
\name{ecdb-class}
\alias{ecdb-class}
\title{setClass for ecdb class}
\description{
setClass for ecdb class
}
\section{Slots}{

\describe{
\item{\code{call}}{the match.call slot}

\item{\code{file}}{character, the full path to an elliptic database.}

\item{\code{conn}}{an object of SQLiteConnection class.}

\item{\code{is.internal}}{logical, whether the connected db is internal.}

\item{\code{conf}}{list of configuration for data generation assigned by the constructor.
Typical user should not have to modify this list 
unless you need to generate more data for advanced research.}
}}

\author{
Stephen H-T. Lihn
}
\keyword{ecdb}
