% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdb-dbSendQuery-method.R
\name{ecdb.dbSendQuery}
\alias{ecdb.dbSendQuery}
\title{Send query to the elliptic database}
\usage{
ecdb.dbSendQuery(db, statement, ...)
}
\arguments{
\item{db}{an object of ecdb class}

\item{statement}{character, the SQL statement}

\item{...}{database-specific parameters may be specified here}
}
\value{
a result set object
}
\description{
This API is used for write operations such as CREATE and INSERT.
}
\author{
Stephen H-T. Lihn
}
\keyword{ecdb}
