% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdb-protectiveCommit-method.R
\name{ecdb.protectiveCommit}
\alias{ecdb.protectiveCommit}
\title{Protective commit}
\usage{
ecdb.protectiveCommit(db)
}
\arguments{
\item{db}{an object of ecdb class}
}
\value{
The db object
}
\description{
Protective commit after sending query to the elliptic database.
}
\author{
Stephen H-T. Lihn
}
\keyword{ecdb}

