% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdq-class.R
\docType{class}
\name{ecdq-class}
\alias{ecdq-class}
\title{setClass for ecdq class}
\description{
setClass for ecdq class, the quantile generator
}
\section{Slots}{

\describe{
\item{\code{call}}{the match.call slot}

\item{\code{xseg.from,xseg.to}}{numeric vectors. The from and to for each x segment.}

\item{\code{cseg.from,cseg.to}}{numeric vectors. The from and to for each cdf segment.}

\item{\code{cseg.min,cseg.max}}{numeric. The min and max of cdf segments.}

\item{\code{N_seg}}{numeric. Number of segments.}

\item{\code{cdf.fit}}{A vector of \code{lm} object, one for each segment.}

\item{\code{x_left_tail,x_right_tail}}{numeric. The starting x of left and right tails.}

\item{\code{fit.left,fit.right}}{objects of \code{lm} class for fitting the tails.}

\item{\code{conf}}{list of miscelaneous configurations. For debugging purpose.}
}}

\keyword{class}
\keyword{constructor}
\keyword{ecdq}
