% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdq-constructor.R
\name{ecdq}
\alias{ecdq}
\title{Constructor of ecdq class}
\usage{
ecdq(ecd, verbose = FALSE)
}
\arguments{
\item{ecd}{An object of ecd class}

\item{verbose}{logical, display timing information, for debugging purpose.}
}
\value{
An object of ecdq class
}
\description{
Construct an ecdq class by providing the required parameters.
}
\examples{
\dontrun{
    d <- ecd()
    dq <- ecdq(d)
}
}
\author{
Stephen H. Lihn
}
\keyword{constructor}
