% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecld-const-method.R
\name{ecld.const}
\alias{ecld.const}
\title{Analytic solution of the normalization constant for lambda distribution}
\usage{
ecld.const(object)
}
\arguments{
\item{object}{an object of ecld class}
}
\value{
numeric
}
\description{
The normalization constant \eqn{C}. SGED is supported.
}
\examples{
ld <- ecld(3)
ecld.const(ld)
}
\author{
Stephen H. Lihn
}
\keyword{ecld}

