% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecld-fixed-point-RN0-method.R
\name{ecld.fixed_point_SN0_atm_ki}
\alias{ecld.fixed_point_SN0_atm_ki}
\alias{ecld.fixed_point_SN0_atm_ki_sd}
\alias{ecld.fixed_point_SN0_lambda_skew_ratio}
\alias{ecld.fixed_point_SN0_rho_sd}
\alias{ecld.fixed_point_SN0_skew}
\title{The ATM RNO related constants and calculations in fixed point model}
\usage{
ecld.fixed_point_SN0_atm_ki(lambda)

ecld.fixed_point_SN0_rho_sd(lambda)

ecld.fixed_point_SN0_atm_ki_sd()

ecld.fixed_point_SN0_skew(lambda, atm_ki = NULL)

ecld.fixed_point_SN0_lambda_skew_ratio(lambda, atm_ki = NULL)
}
\arguments{
\item{lambda}{numeric the lambda parameter.}

\item{atm_ki}{numeric optional and experimental, use it as override.
This is for experimental purpose, default is \code{NULL}.
A typical override is the \code{sd/sigma}.}
}
\value{
numeric
}
\description{
Computes the small sigma limit of ATM location, rho/stdev,
ATM skew of \eqn{Q_c}, and the ratio of lambda to ATM skew 
under the RNO measure in the fixed point model.
}
\author{
Stephen H-T. Lihn
}
\keyword{ATM}

