% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecld-mpnum-method.R
\name{ecld.mpnum}
\alias{ecld.mpnum}
\alias{ecld.ifelse}
\alias{ecld.sapply}
\alias{ecld.mclapply}
\title{Wrappers for ecld to maintain consistent type between mpfr and numeric}
\usage{
ecld.mpnum(object, x)

ecld.ifelse(object, test, yes, no)

ecld.sapply(object, x, FUN, ...)

ecld.mclapply(object, x, FUN, ...)
}
\arguments{
\item{object}{an object of ecd class}

\item{x}{a vector of numeric or mpfr.}

\item{test}{logical, test of \code{ifelse}.}

\item{yes}{return values for true elements of test}

\item{no}{return values for false elements of test}

\item{FUN}{the function to be applied to each element of x}

\item{...}{optional arguments to \code{FUN}}
}
\value{
a numeric or mpfr vector
}
\description{
Primarily to make sure \code{x} is converted to mpfr vector if it is not,
when \code{use.mpfr} is set.
}
\author{
Stephen H-T. Lihn
}
\keyword{utility}
