% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecld.quartic-RN0-method.R
\name{ecld.quartic_SN0_atm_ki}
\alias{ecld.quartic_SN0_atm_ki}
\alias{ecld.quartic_SN0_max_RNV}
\alias{ecld.quartic_SN0_rho_stdev}
\alias{ecld.quartic_SN0_skew}
\title{The ATM RNO related constants and calculations in quartic model}
\usage{
ecld.quartic_SN0_atm_ki()

ecld.quartic_SN0_rho_stdev()

ecld.quartic_SN0_skew()

ecld.quartic_SN0_max_RNV(sigma = 0)
}
\arguments{
\item{sigma}{numeric, the volatility parameter}
}
\value{
numeric
}
\description{
Computes the small sigma limit of ATM location, rho/stdev,
and ATM skew of \eqn{Q_p} under the RNO measure in quartic model.
Computes the maximum risk-neutral violation as an extension of RN0 measure.
}
\author{
Stephen H-T. Lihn
}
\keyword{ATM}

