% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-ecldOrEcd-class.R
\docType{class}
\name{ecldOrEcd-class}
\alias{ecldOrEcd-class}
\title{The ecldOrEcd class}
\description{
The S4 class union of ecld and ecd, primarily used to define slot in \code{ecop.opt} class.
Its usage is rather cumbersome, so the end user should avoid it as much as possible.
}
