% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecop-class.R
\docType{class}
\name{ecop-class}
\alias{ecop-class}
\title{An S4 class to represent the top-level option model}
\description{
The \code{ecop} class serves as an object-oriented container for
the option pricing model. It does have a specific purpose at the moment -
that is, to produce all the data for the charts of the paper, based on
CBOE data structure. Therefore, user may not find it general enough.
That probably will be the case for the time being until more popularity calls
for a more generic container.
}
\section{Slots}{

\describe{
\item{\code{call}}{the match.call slot}

\item{\code{conf}}{list, configuration}

\item{\code{key}}{character}

\item{\code{symbol}}{character}

\item{\code{datadate}}{Date}

\item{\code{days}}{numeric, days between datadate and expiry date}

\item{\code{ttm}}{numeric, time to maturity in days/365}

\item{\code{int_rate}}{numeric}

\item{\code{div_yield}}{numeric}

\item{\code{put_data}}{the put data of ecop.opt class}

\item{\code{call_data}}{the call data of ecop.opt class}

\item{\code{put_conf}}{list, the put configuration}

\item{\code{call_conf}}{list, the call configuration}
}}
\author{
Stephen H-T. Lihn
}
\keyword{ecop}

