% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecop-read-csv-by-symbol.R
\name{ecop.read_csv_by_symbol}
\alias{ecop.read_csv_by_symbol}
\title{Read option data csv}
\usage{
ecop.read_csv_by_symbol(symbol)
}
\arguments{
\item{symbol}{character, option data symbol}
}
\value{
dataframe
}
\description{
Read option data csv into dataframe, enriched with
\code{Date, expiration_date, days}.
}
\examples{

df <- ecop.read_csv_by_symbol("spxoption2")

}
\author{
Stephen H-T. Lihn
}
\keyword{data}

