% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecop-vix-plot-3x3-method.R
\name{ecop.vix_plot_3x3}
\alias{ecop.vix_plot_3x3}
\title{Produce 3x3 plot of VIX volatility smiles for a date}
\usage{
ecop.vix_plot_3x3(date_str, option_data, result, result_avg)
}
\arguments{
\item{date_str}{character in the form of YYYY-MM-DD}

\item{option_data}{dataframe, read from \code{ecop.read_csv_by_symbol}}

\item{result}{dataframe, the VIX optimx result}

\item{result_avg}{dataframe, the VIX optimx result using average lambda for all expirations}
}
\value{
The 3x3 plot
}
\description{
This utlity produces 3x3 plot of volatility smiles for a date. It is used for the VIX option paper.
}
\keyword{structure}
\keyword{term}
