% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdb-history-generic.R
\name{history.ecdb}
\alias{history.ecdb}
\alias{history}
\alias{history,ecdb-method}
\title{List of history in the Elliptic DB}
\usage{
\method{history}{ecdb}(object)

history(object)

\S4method{history}{ecdb}(object)
}
\arguments{
\item{object}{an object of ecdb class.}
}
\value{
list of history
}
\description{
List of unique history reflecting the bootstrap activities.
}
\author{
Stephen H-T. Lihn
}
\keyword{ecdb}
