% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-jinv-generic.R
\docType{methods}
\name{jinv.ecd}
\alias{jinv.ecd}
\alias{jinv}
\alias{jinv,ecd-method}
\title{J-invariant of the elliptic curve \eqn{y(x)}}
\usage{
\method{jinv}{ecd}(object, no.validate = FALSE)

jinv(object, no.validate = FALSE)

\S4method{jinv}{ecd}(object, no.validate = FALSE)
}
\arguments{
\item{object}{an object of ecd class}

\item{no.validate}{logical, if \code{TRUE}, don't validate presence of \code{beta}.
Default is \code{FALSE}.}
}
\value{
the j-invariant
}
\description{
J-invariant of the elliptic curve \eqn{y(x)}
}
\examples{
d <- ecd(1,1)
j <- jinv(d)
}
\author{
Stephen H-T. Lihn
}
\keyword{stats}
