% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp-k2mnt-method.R
\name{k2mnt}
\alias{k2mnt}
\alias{mnt2k}
\title{Conversion between cumulants and moments}
\usage{
k2mnt(k)

mnt2k(m)
}
\arguments{
\item{k}{numeric, first four cumulants.}

\item{m}{numeric, first four moments.}
}
\value{
numeric
}
\description{
Implements conversion between the first four cumulants and moments
}
\author{
Stephen H-T. Lihn
}
\keyword{moments}
