% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp-generate-tau-method.R
\name{lamp.generate_tau}
\alias{lamp.generate_tau}
\title{Generate tau from stable distribution}
\usage{
lamp.generate_tau(object)
}
\arguments{
\item{object}{an object of lamp class}
}
\value{
an object of lamp class with tau populated, tau_i is set to 1.
}
\description{
Generate tau, a random sequence representing the stable random walk process.
}
\examples{
lp <- lamp(4, rnd.n=10)
lp1 <- lamp.generate_tau(lp)
lp1@tau

}
\author{
Stephen H-T. Lihn
}
\keyword{simulation}
