% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp-plot-sim4-method.R
\name{lamp.plot_sim4}
\alias{lamp.plot_sim4}
\alias{lamp.plot_sim6}
\title{Plot the simulation result in standard layout}
\usage{
lamp.plot_sim4(object)

lamp.plot_sim6(object)
}
\arguments{
\item{object}{an object of lamp class}
}
\value{
an object of lamp class
}
\description{
Plot the simulation result in standard layout, with 4 or 6 charts
The PDF and log(PDF) histogram of Z, the lambda process.
The log(PDF) histogram of N, the stable count process.
The log(PDF) histogram of B, the binomial random walk process.
The 6-chart plot also includes the asymptotic kurtosis and stdev 
vs the bps of data points dropped in Z.
}
\author{
Stephen H-T. Lihn
}
\keyword{plot}
