% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp-qsl-fit-config-method.R
\name{lamp.qsl_fit_config}
\alias{lamp.qsl_fit_config}
\alias{lamp.qsl_fit_config_xtable}
\title{Read QSLD fit config}
\usage{
lamp.qsl_fit_config(key = NULL, extdata_dir = NULL, filename = NULL)

lamp.qsl_fit_config_xtable(df)
}
\arguments{
\item{key}{character, the top-level key for config, default to NULL.}

\item{extdata_dir}{optionally specify user's own extdata folder, default is NULL.}

\item{filename}{character, optionally specify user's own config file name, default is NULL.}

\item{df}{the data frame generated from \code{lamp.qsl_fit_config}.}
}
\value{
The data.frame object for the config
}
\description{
Read QSLD fit config for plot or custom fit utility.
The xtable print utility is also provided to
generate high quality latex output for publication purpose.
}
\examples{
c <- lamp.qsl_fit_config()

}
\keyword{data}
\keyword{sample}
