% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp-sd-factor-method.R
\name{lamp.sd_factor}
\alias{lamp.sd_factor}
\title{Calculate sd adjustment factor}
\usage{
lamp.sd_factor(object)
}
\arguments{
\item{object}{an object of lamp class}
}
\value{
numeric, the sd factor
}
\description{
Calculate sd adjustment factor. 
For L2 random walk, it is the power of 1/(1+alpha/2).
For L1 random walk, it is the power of 1.
This factor can be used to adjust either the scale parameter of the stable distribution
or T.inf that cuts off the Levy sums.
}
\author{
Stephen H-T. Lihn
}
\keyword{simulation}
