% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp-stable-rnd-walk-method.R
\name{lamp.stable_rnd_walk}
\alias{lamp.stable_rnd_walk}
\title{Calculate the stable random walk}
\usage{
lamp.stable_rnd_walk(object, n, b)
}
\arguments{
\item{object}{an object of lamp class}

\item{n}{numeric, number of items in Levy sums}

\item{b}{numeric, cumulative sum of signs in Levy sums}
}
\value{
numeric, the value of the random walk
}
\description{
Calculate the stable random walk. There are 4 types of random walk you can specify:
11. Laplace(0,1). No skewess.
1. Experimental Laplace random walk via Gauss-Laplace transmutation.
22. Normal distribution N(0, sqrt(n))*epsilon. No skewess.
2. Binomial random walk, b*epsilon. This can produce skewness.
}
\author{
Stephen H-T. Lihn
}
\keyword{simulation}
