% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levy-dlambda-method.R
\name{levy.dlambda}
\alias{levy.dlambda}
\title{Standard Lambda distribution}
\usage{
levy.dlambda(x, lambda = 4)
}
\arguments{
\item{x}{numeric, complex, mpfr, mpfc}

\item{lambda}{numeric. Default is 4, the quartic distribution.}
}
\value{
PDF in the same type as x
}
\description{
Standard Lambda distribution PDF that can take complex argument.
}
\examples{
x = seq(1,10)
y = levy.dlambda(x)

}
\author{
Stephen H. Lihn
}
\keyword{PDF}
