% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levy-dskewed-distribution-method.R
\name{levy.dskewed}
\alias{levy.dskewed}
\title{Skewed Levy distribution in Levy statistics}
\usage{
levy.dskewed(x, lambda = 4)
}
\arguments{
\item{x}{numeric, complex, mpfr, mpfc}

\item{lambda}{numeric. Default is 4, the Levy distribution as is generally called.}
}
\value{
PDF in the same type as x
}
\description{
Skewed Levy distribution PDF. In our context, "skewed" means "completed asymmetric alpha-stable",
or called "one-sided alpha-stable". And we use lambda = 2/alpha.
}
\examples{
x = seq(1,10)
y = levy.dskewed(x)

}
\author{
Stephen H. Lihn
}
\keyword{PDF}
