% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-quantilize-generic.R
\docType{methods}
\name{quantilize.ecd}
\alias{quantilize}
\alias{quantilize,ecd-method}
\alias{quantilize.ecd}
\title{Add the quantile data to the ecd object}
\usage{
\method{quantilize}{ecd}(object, show.warning = FALSE)

quantilize(object, show.warning = FALSE)

\S4method{quantilize}{ecd}(object, show.warning = FALSE)
}
\arguments{
\item{object}{an object of ecd class}

\item{show.warning}{logical, if \code{TRUE}, display a warning message.
Default is \code{FALSE}.}
}
\value{
an object of ecd class with a newly generated ecdq object.
}
\description{
Add the quantile data to the ecd object if it is not created yet.
}
\examples{
\dontrun{
   d <- ecd(-1,1)
   quantilize(d)
}
}
\author{
Stephen H-T. Lihn
}
\keyword{distribution}

