% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamp-laplace-distribution-method.R
\name{rlaplace0}
\alias{rlaplace0}
\alias{dlaplace0}
\title{Laplace distribution}
\usage{
rlaplace0(n, b = 1)

dlaplace0(x, b = 1)
}
\arguments{
\item{n}{numeric, number of observations.}

\item{b}{numeric, the scale parameter, where the variance is 2*b^2.}

\item{x}{numeric, vector of responses.}
}
\value{
numeric, standard convention is followed: 
        d* returns the density, 
        p* returns the distribution function, 
        q* returns the quantile function, and 
        r* generates random deviates.
}
\description{
Implements some aspects of Laplace distribution (based on stats package) for 
stable random walk simulation.
}
\author{
Stephen H-T. Lihn
}
\keyword{Laplace}
