% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sld-constructor.R
\name{sld}
\alias{sld}
\alias{qsld}
\title{Constructor of sld class}
\usage{
sld(t = 1, nu0 = 0, theta = 1, convo = 1, beta.a = 0, mu = 0, lambda = 4)

qsld(t = 1, nu0 = 0, theta = 1, convo = 1, beta.a = 0, mu = 0)
}
\arguments{
\item{t}{numeric, the time parameter. Must be positive. Default: 1.}

\item{nu0}{numeric, the floor volatility parameter. Must be positive. Default: 0.}

\item{theta}{numeric, the volatility scale parameter. Must be positive. Default: 1.}

\item{convo}{numeric, the convolution parameter. Must be positive. Default: 1.}

\item{beta.a}{numeric, the skewness parameter. Default: 0.}

\item{mu}{numeric, the location parameter. Default: 0.}

\item{lambda}{numeric, the lambda parameter. Must be positive. Default: 4.}
}
\value{
an object of sld class
}
\description{
Construct an \code{\link{sld-class}} by providing the required parameters.
The \code{qsld} constructor is also provided by hiding and defaulting lambda to 4.
}
\examples{
d <- sld()
d <- qsld()

}
\author{
Stephen H-T. Lihn
}
\keyword{constructor}
