% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-solve-sym-generic.R
\docType{methods}
\name{solve_sym.ecd}
\alias{solve_sym}
\alias{solve_sym,ecd-method}
\alias{solve_sym.ecd}
\title{Analytic solution for a symmetric elliptic curve}
\usage{
\method{solve_sym}{ecd}(object, x)

solve_sym(object, x)

\S4method{solve_sym}{ecd}(object, x)
}
\arguments{
\item{object}{an object of ecd class}

\item{x}{array of x dimension}
}
\value{
array of y
}
\description{
Analytic solution for a symmetric elliptic curve \eqn{y(x)}
}
\examples{
d <- ecd()
x <- seq(-100,100,by=0.01)
y <- solve_sym(d,x)
}
\author{
Stephen H-T. Lihn
}
\keyword{solve}

