% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdb-summary-generic.R
\name{summary.ecdb}
\alias{summary.ecdb}
\alias{summary}
\alias{summary,ecdb-method}
\title{Summary for the Elliptic DB (ECDB)}
\usage{
\method{summary}{ecdb}(object, ...)

summary(object, ...)

\S4method{summary}{ecdb}(object, ...)
}
\arguments{
\item{object}{an object of ecdb class.}

\item{...}{more arguments for summary. Currently not used.}
}
\description{
Summary for the Elliptic DB (ECDB)
}
\examples{
summary(ecdb())
}
\author{
Stephen H-T. Lihn
}
\keyword{ecdb}
