% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdb-write-generic.R
\docType{methods}
\name{write.ecdb}
\alias{write}
\alias{write,list,ecdb-method}
\alias{write.ecdb}
\title{Write API for the ecdb for a list of basic ecdattr objects}
\usage{
\method{write}{ecdb}(x, object)

write(x, object)

\S4method{write}{list,ecdb}(x, object)
}
\arguments{
\item{x}{a list of basic ecdattr objects}

\item{object}{an object of ecdb class}
}
\value{
The row count
}
\description{
It takes a list of basic ecdattr objects, enrich them in parallel, then save them to ecdb.
}
\keyword{ecdb}

