% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-y-slope-generic.R
\docType{methods}
\name{y_slope.ecd}
\alias{y_slope}
\alias{y_slope,ecd-method}
\alias{y_slope.ecd}
\title{Slope of \eqn{y(x)}}
\usage{
\method{y_slope}{ecd}(object, x)

y_slope(object, x)

\S4method{y_slope}{ecd}(object, x)
}
\arguments{
\item{object}{an object of ecd class}

\item{x}{a numeric vector of \code{x} dimension}
}
\value{
a numeric vector of \eqn{dy/dx}
}
\description{
Slope of \eqn{y(x)}, that is, \eqn{dy/dx}.
}
\examples{
d <- ecd(0,1)
x <- seq(-20,20,by=0.01)
yp <- y_slope(d,x)
}
\author{
Stephen H. Lihn
}
\keyword{elliptic-curve}

