% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-package.R
\docType{package}
\name{ecd-package}
\alias{ecd-package}
\title{ecd: A package for the elliptic distribution.}
\description{
The ecd package provides the core class and functions to calculate 
the elliptic distribution. They are either generic or use \code{ecd.} namespace.
The lambda distribution is using \code{ecld.} namespace. SGED is considered part of ecld.
The option pricing API is using \code{ecop.} namespace.
}
\author{
Stephen H-T. Lihn
}

