% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd-rational-method.R
\name{ecd.rational}
\alias{ecd.rational}
\title{Utility to convert a numeric to a rational}
\usage{
ecd.rational(x)
}
\arguments{
\item{x}{numeric}
}
\value{
a vector of two integers, representing numerator and denominator
}
\description{
Convert a numeric x to rational p/q, which is then used for polynomial construction
}
\examples{
pq <- ecd.rational(2.5)
}
\keyword{solve}

