% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecop-read-csv-by-symbol.R
\name{ecop.read_csv_by_symbol}
\alias{ecop.read_csv_by_symbol}
\alias{ecop.enrich_option_df}
\title{Read option data csv}
\usage{
ecop.read_csv_by_symbol(symbol, extdata_dir = NULL)

ecop.enrich_option_df(df)
}
\arguments{
\item{symbol}{character, option data symbol}

\item{extdata_dir}{optionally specify user's own extdata folder}

\item{df}{dataframe, it is assumed to be in CBOE heading format}
}
\value{
dataframe
}
\description{
Read option data csv into dataframe. The dataframe is enriched with
\code{Date, expiration_date, days}.
}
\examples{

df <- ecop.read_csv_by_symbol("spxoption2")

}
\author{
Stephen H-T. Lihn
}
\keyword{data}
